requirements.txt
python-telegram-bot==21.10
paramiko==3.5.1
python-dotenv==1.0.1
pytest==8.3.5
